+++ compile-shaders.sh
@@ -54,6 +54,7 @@ compile attention_bundled            attention.comp        vulkan1.1 SUBGROUPS H
 compile attention_bundle_exact       attention.comp        vulkan1.1 SUBGROUPS WIDE FOURS GROUPED
 compile combine                      combine.comp          vulkan1.0
 compile row_product                  row_product.comp      vulkan1.0
+compile row_product_low              row_product.comp      vulkan1.0 LOW_BITS
 compile row_product_super            row_product_super.comp vulkan1.1 NUM_ROWS=2u
 compile row_product_super5           row_product_super5.comp vulkan1.1 NUM_ROWS=2u
 compile row_product_super6           row_product_super6.comp vulkan1.1 NUM_ROWS=2u